+++ test/support/image_test_helpers.ex
@@ -4,7 +4,7 @@ defmodule Image.TestSupport do
 
   @images_path Path.join(__DIR__, "images")
   @validate_path Path.join(__DIR__, "validate")
-  @acceptible_similarity 0.5
+  @acceptible_similarity 1.1
 
   def assert_files_equal(expected, result) do
     assert File.read!(expected) == File.read!(result)